@TomFinley 2019-02-19
Hmmm. I wasn't aware BlockingCollection was so problematic.