@stephentoub 2019-02-19
> I wasn't aware BlockingCollection was so problematic.

It shouldn't be; this is the first time I've seen this particular issue, mainly because it's the first time I've seen an app create 15,000 BlockingCollections :-)